requirements.txt
asyncpg
discord.py
PyGithub
DiscordBotsOrgAPI